SMAR-6: Amélioration de la gestion des permissions pour les opération…

+++ src/main/java/com/example/demo/controller/OrderController.java
@@ -1,8 +1,11 @@
 package com.example.demo.controller;
 
 import com.example.demo.dto.OrderDTO;
+import com.example.demo.enums.Role;
 import com.example.demo.exception.ApiResponse;
+import com.example.demo.service.HelperService;
 import com.example.demo.service.OrderService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -16,33 +19,80 @@ import java.util.List;
 public class OrderController {
 
     private final OrderService orderService;
+    private final HelperService helperService ;
 
     @PostMapping
-    public ResponseEntity<OrderDTO> createOrder(@RequestBody OrderDTO orderDTO) {
+    public ResponseEntity<?> createOrder(@RequestBody OrderDTO orderDTO , HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         OrderDTO createdOrder = orderService.create(orderDTO);
         return ResponseEntity.ok(createdOrder);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<OrderDTO> updateOrder(@PathVariable Long id, @RequestBody OrderDTO orderDTO) {
+    public ResponseEntity<?> updateOrder(@PathVariable Long id, @RequestBody OrderDTO orderDTO , HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
+
         OrderDTO updatedOrder = orderService.update(id, orderDTO);
         return ResponseEntity.ok(updatedOrder);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<OrderDTO> getOrderById(@PathVariable Long id) {
+    public ResponseEntity<?> getOrderById(@PathVariable Long id , HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         OrderDTO order = orderService.getById(id);
         return ResponseEntity.ok(order);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse> deleteOrder(@PathVariable Long id) {
+    public ResponseEntity<?> deleteOrder(@PathVariable Long id , HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         ApiResponse response = orderService.delete(id);
         return ResponseEntity.ok(response);
     }
 
     @GetMapping
-    public ResponseEntity<List<OrderDTO>> getAllOrders() {
+    public ResponseEntity<?> getAllOrders(HttpServletRequest httpRequest) {
+
+        if(!helperService.isAuthenticated(httpRequest)) {
+            return ResponseEntity.status(401).body("tu ne pas login");
+        }
+
+        if (!helperService.hasRole(httpRequest, Role.ADMIN)) {
+            return ResponseEntity.status(403).body("vous n'avez pas la permission pour faire cette action : ADMIN only");
+        }
+
         List<OrderDTO> orders = orderService.getAll();
         return ResponseEntity.ok(orders);
     }
